requirements.txt
absl-py==1.4.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dm-control==1.0.10
dm-env==1.6
dm-tree==0.1.8
fonttools==4.38.0
glfw==2.5.6
gym==0.21.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
kiwisolver==1.4.4
labmaze==1.0.6
lxml==4.9.2
matplotlib==3.7.0
mujoco==2.3.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.0
PyOpenGL==3.1.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scipy==1.10.1
six==1.16.0
stable-baselines3==1.7.0
torch==1.13.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.14.0
